Fix is_a_git_repository()
A typo in argument was preventing it to work properly, and logic was flawed.

If target path was a directory inside another git repository,
this function was returning True, because of the use of `--is-inside-work-tree`.

So, instead, get the .git directory if any, but compare its path to the actual target directory.